refactor to use same mock docker client
